+++ Specs/Scene/ModelExperimental/loadAndZoomToModelExperimental.js
@@ -23,6 +23,7 @@ function loadAndZoomToModelExperimental(options, scene) {
         featureIdLabel: options.featureIdLabel,
         instanceFeatureIdLabel: options.instanceFeatureIdLabel,
         incrementallyLoadTextures: options.incrementallyLoadTextures,
+        clippingPlanes: options.clippingPlanes,
         lightColor: options.lightColor,
         imageBasedLighting: options.imageBasedLighting,
         backFaceCulling: options.backFaceCulling,