+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 1.93 - 2022-05-02
+
+##### Additions :tada:
+- Added experimental voxel rendering that supports glTF with EXT_primitive_voxels extension, 3D Tiles, and procedural data. [#10253](https://github.com/CesiumGS/cesium/pull/10253).
+
 ### 1.92 - 2022-04-01
 
 ##### Breaking Changes :mega:
@@ -21,7 +26,7 @@
 - Refactored metadata API so `tileset.metadata` and `content.group.metadata` are more symmetric with `content.metadata` and `tile.metadata`. [#10224](https://github.com/CesiumGS/cesium/pull/10224)
 - Added support for `EXT_structural_metadata` property attributes in `CustomShader` [#10228](https://github.com/CesiumGS/cesium/pull/10228)
 - Added partial support for `EXT_structural_metadata` property textures in `CustomShader` [#10247](https://github.com/CesiumGS/cesium/pull/10247)
-- Added experimental voxel rendering that supports glTF with EXT_primitive_voxels, 3D Tiles, and procedural data. [#10253](https://github.com/CesiumGS/cesium/pull/10253).
+- Added clipping planes to `ModelExperimental`. [#10250](https://github.com/CesiumGS/cesium/pull/10250)
 
 ##### Fixes :wrench: